Triangle locking + Adam tuning
* If the links are all free, there is a large ambiguity because one can
trade off the M link with the lower H link, so every time it trains to
a different set of couplings, there is no stable solution, initial
condition sensitive. To overcome this difficulty, we lock the M link to
its lower H link, i.e. in the dual lattice, the vertical link is locked
to the horizontal link in the IR layer (not in the UV layer), in this
way the external field is still free to tune, such that the first layer
can still learn the feature, and there is no ambiguity now. Every time
the machine converge to the unique minimum (locally stable over a large
parameter regime).
* To prevent the coupling becoming negative in the free case, we also
introduce the wall potential on the negative side. Although triangle
locking solves the problem, but we still keep the wall, because it is
of little computation cost and does not affect the fixed point.
* Adam parameter is also tuned to fit this model.
TODO:
* We observe fast convergence at the beginning follow by very slow
tail, during the tail training, the machine is adjusting the external
field and the couplings correspondingly, as the field gets greater than
2~3, its gradient is very small, so what about fixing the field to ln D?
* What about multiple layer on square lattice, and on hyperbolic
lattice?